Refactor: Split main() -> load_config(), setup() and main()
